+++ db/migrate/20141119011011_add_callback_params_to_asyncapi_client_jobs.rb
@@ -1,0 +1,5 @@
+class AddCallbackParamsToAsyncapiClientJobs < ActiveRecord::Migration
+  def change
+    add_column :asyncapi_client_jobs, :callback_params, :text
+  end
+end