+++ spec/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20141104030959) do
+ActiveRecord::Schema.define(version: 20141119011011) do
 
   create_table "asyncapi_client_jobs", force: true do |t|
     t.string   "server_job_url"
@@ -26,6 +26,7 @@ ActiveRecord::Schema.define(version: 20141104030959) do
     t.text     "headers"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.text     "callback_params"
   end
 
 end